+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## [Unreleased]
+## 2.0.0 - 2023-09-17
+
+### Added
+
+- Support for `innmind/immutable:~5.0`
+
+### Changed
+
+- All encodings are represented with `Innmind\Immutable\Str\Encoding`
 
 ### Removed